Implementation of the BAS algorithm for optimization problems.

You could find more information about BAS in https://arxiv.org/abs/1710.10724.

BASoptim(fn, init = NULL, lower = c(-6, 0), upper = c(-1, 2),
  d0 = 0.001, d1 = 3, eta_d = 0.95, l0 = 0, l1 = 0,
  eta_l = 0.95, step = 0.8, eta_step = 0.95, n = 200,
  seed = NULL, trace = T, steptol = 0.01)

Arguments

fn
objective function; function need to be optimized

init
default = NULL, it will generate randomly; Of course, you can specify it.

lower
lower of parameters to be estimated; Default = c(-6,0) because of the test on Michalewicz function of which thelower is c(-6,0); By the way, you should set one of init or lower parameter at least to make the code know the dimensionality of your problem.

upper
upper of parameters; Default = c(-1,2).

d0
a constant to gurantee that sensing length of antennae d doesn't equal to zero. More specifically, $$d^t = \eta_d * d^{t-1} + d_0$$where attenuation coefficient \(\eta_d\) belongs to \([0,1]\)

d1
initial value of antenae length. You can specify it according to your problem scale

eta_d
attenuation coefficient of sensing length of antennae

l0
position jitter factor constant.Default = 0.

l1
initial position jitter factor.Default = 0.$$x = x - step * dir * sign(fn(left) - fn(right)) + l *random(npars)$$

eta_l
attenuation coefficient of jitter factor.$$l^t = \eta_l * l^{t-1} + l_0$$

step
initial step-size of beetle

eta_step
attenuation coefficient of step-size.$$step^t = \eta_step * step^{t-1}$$

n
iterations times

seed
random seed; default = NULL ; The results of BAS depend on random init value and random directions. Therefore, if you set a random seed, for example,seed = 1, the results will remain the same no matter how many times you repeat your experiments.

trace
default = T; trace the process of BAS iteration.

steptol
default = 0.01; Iteration will stop if step-size in current moment is less than steptol.

Value

A list including best beetle position (parameters) and corresponding objective function value.

References

X. Y. Jiang, and S. Li, BAS: beetle antennae search algorithm for optimization problems, arXiv:1710.10724v1.

Examples

#======== examples start =======================
# BAS application on Michalewicz function
library(rBAS)
mich <- function(x){
  y1 <- -sin(x[1])*(sin((x[1]^2)/pi))^20
  y2 <- -sin(x[2])*(sin((2*x[2]^2)/pi))^20
  return(y1+y2)
}
BASoptim(fn = mich,
         lower = c(-6,0), upper = c(-1,2),
         seed = 1, n = 100)

#> Iter:  0  xbest:  [ -4.672457 0.7442478 ], fbest=  -6.583555e-05 
#> Iter:  1  xbest:  [ -4.813013 0 ], fbest=  -0.09012505 
#> Iter:  2  xbest:  [ -4.813013 0 ], fbest=  -0.09012505 
#> Iter:  3  xbest:  [ -5.042063 0.5342482 ], fbest=  -0.5335012 
#> Iter:  4  xbest:  [ -5.042063 0.5342482 ], fbest=  -0.5335012 
#> Iter:  5  xbest:  [ -5.042063 0.5342482 ], fbest=  -0.5335012 
#> Iter:  6  xbest:  [ -5.042063 0.5342482 ], fbest=  -0.5335012 
#> Iter:  7  xbest:  [ -5.042063 0.5342482 ], fbest=  -0.5335012 
#> Iter:  8  xbest:  [ -5.042063 0.5342482 ], fbest=  -0.5335012 
#> Iter:  9  xbest:  [ -4.809573 1.495284 ], fbest=  -0.882469 
#> Iter:  10  xbest:  [ -4.809573 1.495284 ], fbest=  -0.882469 
#> Iter:  11  xbest:  [ -3.8701 1.397651 ], fbest=  -0.9768905 
#> Iter:  12  xbest:  [ -3.8701 1.397651 ], fbest=  -0.9768905 
#> Iter:  13  xbest:  [ -3.8701 1.397651 ], fbest=  -0.9768905 
#> Iter:  14  xbest:  [ -3.8701 1.397651 ], fbest=  -0.9768905 
#> Iter:  15  xbest:  [ -3.8701 1.397651 ], fbest=  -0.9768905 
#> Iter:  16  xbest:  [ -3.8701 1.397651 ], fbest=  -0.9768905 
#> Iter:  17  xbest:  [ -3.8701 1.397651 ], fbest=  -0.9768905 
#> Iter:  18  xbest:  [ -3.8701 1.397651 ], fbest=  -0.9768905 
#> Iter:  19  xbest:  [ -3.8701 1.397651 ], fbest=  -0.9768905 
#> Iter:  20  xbest:  [ -3.8701 1.397651 ], fbest=  -0.9768905 
#> Iter:  21  xbest:  [ -3.8701 1.397651 ], fbest=  -0.9768905 
#> Iter:  22  xbest:  [ -3.891725 1.55418 ], fbest=  -1.59477 
#> Iter:  23  xbest:  [ -3.891725 1.55418 ], fbest=  -1.59477 
#> Iter:  24  xbest:  [ -3.891725 1.55418 ], fbest=  -1.59477 
#> Iter:  25  xbest:  [ -3.891725 1.55418 ], fbest=  -1.59477 
#> Iter:  26  xbest:  [ -3.891725 1.55418 ], fbest=  -1.59477 
#> Iter:  27  xbest:  [ -3.891725 1.55418 ], fbest=  -1.59477 
#> Iter:  28  xbest:  [ -3.891725 1.55418 ], fbest=  -1.59477 
#> Iter:  29  xbest:  [ -3.891725 1.55418 ], fbest=  -1.59477 
#> Iter:  30  xbest:  [ -3.891725 1.55418 ], fbest=  -1.59477 
#> Iter:  31  xbest:  [ -3.891725 1.55418 ], fbest=  -1.59477 
#> Iter:  32  xbest:  [ -3.891725 1.55418 ], fbest=  -1.59477 
#> Iter:  33  xbest:  [ -3.891725 1.55418 ], fbest=  -1.59477 
#> Iter:  34  xbest:  [ -3.891725 1.55418 ], fbest=  -1.59477 
#> Iter:  35  xbest:  [ -3.891725 1.55418 ], fbest=  -1.59477 
#> Iter:  36  xbest:  [ -5.028429 1.59817 ], fbest=  -1.619078 
#> Iter:  37  xbest:  [ -4.927904 1.521834 ], fbest=  -1.747204 
#> Iter:  38  xbest:  [ -4.927904 1.521834 ], fbest=  -1.747204 
#> Iter:  39  xbest:  [ -4.927904 1.521834 ], fbest=  -1.747204 
#> Iter:  40  xbest:  [ -4.927904 1.521834 ], fbest=  -1.747204 
#> Iter:  41  xbest:  [ -4.927904 1.521834 ], fbest=  -1.747204 
#> Iter:  42  xbest:  [ -4.927904 1.521834 ], fbest=  -1.747204 
#> Iter:  43  xbest:  [ -4.927904 1.521834 ], fbest=  -1.747204 
#> Iter:  44  xbest:  [ -4.927904 1.521834 ], fbest=  -1.747204 
#> Iter:  45  xbest:  [ -4.960227 1.591408 ], fbest=  -1.947337 
#> Iter:  46  xbest:  [ -4.960227 1.591408 ], fbest=  -1.947337 
#> Iter:  47  xbest:  [ -4.960227 1.591408 ], fbest=  -1.947337 
#> Iter:  48  xbest:  [ -4.960227 1.591408 ], fbest=  -1.947337 
#> Iter:  49  xbest:  [ -4.960227 1.591408 ], fbest=  -1.947337 
#> Iter:  50  xbest:  [ -4.960227 1.591408 ], fbest=  -1.947337 
#> Iter:  51  xbest:  [ -4.960227 1.591408 ], fbest=  -1.947337 
#> Iter:  52  xbest:  [ -4.960227 1.591408 ], fbest=  -1.947337 
#> Iter:  53  xbest:  [ -4.960227 1.591408 ], fbest=  -1.947337 
#> Iter:  54  xbest:  [ -4.960227 1.591408 ], fbest=  -1.947337 
#> Iter:  55  xbest:  [ -4.960227 1.591408 ], fbest=  -1.947337 
#> Iter:  56  xbest:  [ -4.960227 1.591408 ], fbest=  -1.947337 
#> Iter:  57  xbest:  [ -4.960227 1.591408 ], fbest=  -1.947337 
#> Iter:  58  xbest:  [ -4.960227 1.591408 ], fbest=  -1.947337 
#> Iter:  59  xbest:  [ -4.960227 1.591408 ], fbest=  -1.947337 
#> Iter:  60  xbest:  [ -4.973693 1.579025 ], fbest=  -1.959349 
#> Iter:  61  xbest:  [ -4.973693 1.579025 ], fbest=  -1.959349 
#> Iter:  62  xbest:  [ -4.973693 1.579025 ], fbest=  -1.959349 
#> Iter:  63  xbest:  [ -4.973693 1.579025 ], fbest=  -1.959349 
#> Iter:  64  xbest:  [ -4.973693 1.579025 ], fbest=  -1.959349 
#> Iter:  65  xbest:  [ -4.973693 1.579025 ], fbest=  -1.959349 
#> Iter:  66  xbest:  [ -4.973693 1.579025 ], fbest=  -1.959349 
#> Iter:  67  xbest:  [ -4.973693 1.579025 ], fbest=  -1.959349 
#> Iter:  68  xbest:  [ -4.967715 1.583254 ], fbest=  -1.961248 
#> Iter:  69  xbest:  [ -4.967715 1.583254 ], fbest=  -1.961248 
#> Iter:  70  xbest:  [ -4.967715 1.583254 ], fbest=  -1.961248 
#> Iter:  71  xbest:  [ -4.967715 1.583254 ], fbest=  -1.961248 
#> Iter:  72  xbest:  [ -4.967715 1.583254 ], fbest=  -1.961248 
#> Iter:  73  xbest:  [ -4.967715 1.583254 ], fbest=  -1.961248 
#> Iter:  74  xbest:  [ -4.967715 1.583254 ], fbest=  -1.961248 
#> Iter:  75  xbest:  [ -4.972746 1.566607 ], fbest=  -1.962718 
#> Iter:  76  xbest:  [ -4.972746 1.566607 ], fbest=  -1.962718 
#> Iter:  77  xbest:  [ -4.972746 1.566607 ], fbest=  -1.962718 
#> Iter:  78  xbest:  [ -4.96958 1.562077 ], fbest=  -1.963546 
#> Iter:  79  xbest:  [ -4.96958 1.562077 ], fbest=  -1.963546 
#> Iter:  80  xbest:  [ -4.96958 1.562077 ], fbest=  -1.963546 
#> Iter:  81  xbest:  [ -4.970852 1.571799 ], fbest=  -1.965519 
#> Iter:  82  xbest:  [ -4.970852 1.571799 ], fbest=  -1.965519 
#> Iter:  83  xbest:  [ -4.970852 1.571799 ], fbest=  -1.965519 
#> Iter:  84  xbest:  [ -4.970852 1.571799 ], fbest=  -1.965519 
#> Iter:  85  xbest:  [ -4.970852 1.571799 ], fbest=  -1.965519 
#> Iter:  86  xbest:  [ -4.970852 1.571799 ], fbest=  -1.965519 

#> ----step < steptol---------stop the iteration------

#> $par
#> [1] -4.970852  1.571799
#> 
#> $value
#> [1] -1.965519
#> 
#> $df
#> $df$x
#>            [,1]      [,2]
#>  [1,] -4.672457 0.7442478
#>  [2,] -4.813013 0.0000000
#>  [3,] -5.527720 0.0000000
#>  [4,] -5.042063 0.5342482
#>  [5,] -5.727938 0.5400962
#>  [6,] -5.986547 1.1381854
#>  [7,] -5.754145 0.5644425
#>  [8,] -5.182530 0.7025999
#>  [9,] -5.281490 1.2524353
#> [10,] -4.809573 1.4952843
#> [11,] -4.340889 1.3094006
#> [12,] -3.870100 1.3976510
#> [13,] -4.324622 1.3759390
#> [14,] -4.297888 0.9444784
#> [15,] -3.997605 1.2246265
#> [16,] -4.335805 1.4191266
#> [17,] -4.247870 1.7791769
#> [18,] -4.219595 1.4282126
#> [19,] -4.076300 1.1259638
#> [20,] -4.057174 0.8087684
#> [21,] -3.901465 1.0673953
#> [22,] -3.758956 1.3162705
#> [23,] -3.891725 1.5541800
#> [24,] -3.760149 1.7770682
#> [25,] -3.928381 1.5977421
#> [26,] -4.149703 1.5230353
#> [27,] -4.055920 1.3219146
#> [28,] -4.266370 1.3094896
#> [29,] -4.446443 1.2218327
#> [30,] -4.557865 1.0676100
#> [31,] -4.622572 1.2363790
#> [32,] -4.614352 1.4078932
#> [33,] -4.620578 1.5708998
#> [34,] -4.647592 1.7234963
#> [35,] -4.791103 1.6906526
#> [36,] -4.930945 1.6884496
#> [37,] -5.028429 1.5981697
#> [38,] -4.927904 1.5218344
#> [39,] -4.820590 1.5753387
#> [40,] -4.820748 1.6892552
#> [41,] -4.886964 1.6036562
#> [42,] -4.906352 1.7046212
#> [43,] -4.997737 1.6701546
#> [44,] -4.964560 1.7568061
#> [45,] -4.939061 1.6724285
#> [46,] -4.960227 1.5914085
#> [47,] -5.039776 1.5907210
#> [48,] -5.096798 1.5411229
#> [49,] -5.160375 1.5744794
#> [50,] -5.097442 1.5481836
#> [51,] -5.062375 1.4936966
#> [52,] -5.057912 1.4323027
#> [53,] -5.067673 1.4899604
#> [54,] -5.109656 1.5263431
#> [55,] -5.064434 1.4991348
#> [56,] -5.111717 1.5158105
#> [57,] -5.064163 1.5131029
#> [58,] -5.042836 1.5530108
#> [59,] -5.026376 1.5927214
#> [60,] -5.005896 1.5573906
#> [61,] -4.973693 1.5790253
#> [62,] -4.945899 1.5548200
#> [63,] -4.952567 1.5204477
#> [64,] -4.935346 1.5489051
#> [65,] -4.934134 1.5804812
#> [66,] -4.942561 1.6092935
#> [67,] -4.925131 1.5867214
#> [68,] -4.948518 1.6003981
#> [69,] -4.967715 1.5832541
#> [70,] -4.981100 1.5627925
#> [71,] -4.979161 1.5396452
#> [72,] -4.958229 1.5326589
#> [73,] -4.948599 1.5512798
#> [74,] -4.960470 1.5352890
#> [75,] -4.959747 1.5541948
#> [76,] -4.972746 1.5666071
#> [77,] -4.969517 1.5833740
#> [78,] -4.957865 1.5720892
#> [79,] -4.969580 1.5620774
#> [80,] -4.954943 1.5623917
#> [81,] -4.968384 1.5588191
#> [82,] -4.970852 1.5717989
#> [83,] -4.961539 1.5802138
#> [84,] -4.972515 1.5848723
#> [85,] -4.961734 1.5883477
#> [86,] -4.963587 1.5777470
#> [87,] -4.960089 1.5681406
#> 
#> $df$f
#>                [,1]
#>  [1,] -6.583555e-05
#>  [2,] -9.012505e-02
#>  [3,] -1.936756e-11
#>  [4,] -5.335012e-01
#>  [5,] -2.113946e-02
#>  [6,] -5.268663e-02
#>  [7,] -5.829834e-02
#>  [8,] -4.520459e-03
#>  [9,] -2.955197e-02
#> [10,] -8.824690e-01
#> [11,] -8.843190e-02
#> [12,] -9.768905e-01
#> [13,] -2.500279e-01
#> [14,] -3.332599e-06
#> [15,] -1.957705e-01
#> [16,] -4.245035e-01
#> [17,] -1.267595e-01
#> [18,] -4.674003e-01
#> [19,] -2.495379e-02
#> [20,] -4.282447e-02
#> [21,] -5.774847e-01
#> [22,] -4.633610e-01
#> [23,] -1.594770e+00
#> [24,] -5.015277e-01
#> [25,] -1.444227e+00
#> [26,] -9.152899e-01
#> [27,] -1.541317e-01
#> [28,] -8.857080e-02
#> [29,] -1.519483e-02
#> [30,] -2.402070e-04
#> [31,] -2.098338e-02
#> [32,] -3.741238e-01
#> [33,] -1.000000e+00
#> [34,] -3.480996e-01
#> [35,] -5.733793e-01
#> [36,] -1.399919e+00
#> [37,] -1.619078e+00
#> [38,] -1.747204e+00
#> [39,] -1.113107e+00
#> [40,] -6.536957e-01
#> [41,] -1.474686e+00
#> [42,] -1.128607e+00
#> [43,] -1.525702e+00
#> [44,] -1.167127e+00
#> [45,] -1.538486e+00
#> [46,] -1.947337e+00
#> [47,] -1.536824e+00
#> [48,] -1.123031e+00
#> [49,] -1.013604e+00
#> [50,] -1.134311e+00
#> [51,] -1.162911e+00
#> [52,] -8.901292e-01
#> [53,] -1.106272e+00
#> [54,] -1.031182e+00
#> [55,] -1.172251e+00
#> [56,] -9.874940e-01
#> [57,] -1.232900e+00
#> [58,] -1.514348e+00
#> [59,] -1.647125e+00
#> [60,] -1.816259e+00
#> [61,] -1.959349e+00
#> [62,] -1.919300e+00
#> [63,] -1.855705e+00
#> [64,] -1.862077e+00
#> [65,] -1.870589e+00
#> [66,] -1.856385e+00
#> [67,] -1.808746e+00
#> [68,] -1.903151e+00
#> [69,] -1.961248e+00
#> [70,] -1.943276e+00
#> [71,] -1.913278e+00
#> [72,] -1.906078e+00
#> [73,] -1.923817e+00
#> [74,] -1.916147e+00
#> [75,] -1.953076e+00
#> [76,] -1.962718e+00
#> [77,] -1.960208e+00
#> [78,] -1.961383e+00
#> [79,] -1.963546e+00
#> [80,] -1.953223e+00
#> [81,] -1.961547e+00
#> [82,] -1.965519e+00
#> [83,] -1.962314e+00
#> [84,] -1.955660e+00
#> [85,] -1.953549e+00
#> [86,] -1.965323e+00
#> [87,] -1.964181e+00
#> 
#> $df$xbest
#>            [,1]      [,2]
#>  [1,] -4.672457 0.7442478
#>  [2,] -4.813013 0.0000000
#>  [3,] -4.813013 0.0000000
#>  [4,] -5.042063 0.5342482
#>  [5,] -5.042063 0.5342482
#>  [6,] -5.042063 0.5342482
#>  [7,] -5.042063 0.5342482
#>  [8,] -5.042063 0.5342482
#>  [9,] -5.042063 0.5342482
#> [10,] -4.809573 1.4952843
#> [11,] -4.809573 1.4952843
#> [12,] -3.870100 1.3976510
#> [13,] -3.870100 1.3976510
#> [14,] -3.870100 1.3976510
#> [15,] -3.870100 1.3976510
#> [16,] -3.870100 1.3976510
#> [17,] -3.870100 1.3976510
#> [18,] -3.870100 1.3976510
#> [19,] -3.870100 1.3976510
#> [20,] -3.870100 1.3976510
#> [21,] -3.870100 1.3976510
#> [22,] -3.870100 1.3976510
#> [23,] -3.891725 1.5541800
#> [24,] -3.891725 1.5541800
#> [25,] -3.891725 1.5541800
#> [26,] -3.891725 1.5541800
#> [27,] -3.891725 1.5541800
#> [28,] -3.891725 1.5541800
#> [29,] -3.891725 1.5541800
#> [30,] -3.891725 1.5541800
#> [31,] -3.891725 1.5541800
#> [32,] -3.891725 1.5541800
#> [33,] -3.891725 1.5541800
#> [34,] -3.891725 1.5541800
#> [35,] -3.891725 1.5541800
#> [36,] -3.891725 1.5541800
#> [37,] -5.028429 1.5981697
#> [38,] -4.927904 1.5218344
#> [39,] -4.927904 1.5218344
#> [40,] -4.927904 1.5218344
#> [41,] -4.927904 1.5218344
#> [42,] -4.927904 1.5218344
#> [43,] -4.927904 1.5218344
#> [44,] -4.927904 1.5218344
#> [45,] -4.927904 1.5218344
#> [46,] -4.960227 1.5914085
#> [47,] -4.960227 1.5914085
#> [48,] -4.960227 1.5914085
#> [49,] -4.960227 1.5914085
#> [50,] -4.960227 1.5914085
#> [51,] -4.960227 1.5914085
#> [52,] -4.960227 1.5914085
#> [53,] -4.960227 1.5914085
#> [54,] -4.960227 1.5914085
#> [55,] -4.960227 1.5914085
#> [56,] -4.960227 1.5914085
#> [57,] -4.960227 1.5914085
#> [58,] -4.960227 1.5914085
#> [59,] -4.960227 1.5914085
#> [60,] -4.960227 1.5914085
#> [61,] -4.973693 1.5790253
#> [62,] -4.973693 1.5790253
#> [63,] -4.973693 1.5790253
#> [64,] -4.973693 1.5790253
#> [65,] -4.973693 1.5790253
#> [66,] -4.973693 1.5790253
#> [67,] -4.973693 1.5790253
#> [68,] -4.973693 1.5790253
#> [69,] -4.967715 1.5832541
#> [70,] -4.967715 1.5832541
#> [71,] -4.967715 1.5832541
#> [72,] -4.967715 1.5832541
#> [73,] -4.967715 1.5832541
#> [74,] -4.967715 1.5832541
#> [75,] -4.967715 1.5832541
#> [76,] -4.972746 1.5666071
#> [77,] -4.972746 1.5666071
#> [78,] -4.972746 1.5666071
#> [79,] -4.969580 1.5620774
#> [80,] -4.969580 1.5620774
#> [81,] -4.969580 1.5620774
#> [82,] -4.970852 1.5717989
#> [83,] -4.970852 1.5717989
#> [84,] -4.970852 1.5717989
#> [85,] -4.970852 1.5717989
#> [86,] -4.970852 1.5717989
#> [87,] -4.970852 1.5717989
#> 
#> $df$fbest
#>                [,1]
#>  [1,] -6.583555e-05
#>  [2,] -9.012505e-02
#>  [3,] -9.012505e-02
#>  [4,] -5.335012e-01
#>  [5,] -5.335012e-01
#>  [6,] -5.335012e-01
#>  [7,] -5.335012e-01
#>  [8,] -5.335012e-01
#>  [9,] -5.335012e-01
#> [10,] -8.824690e-01
#> [11,] -8.824690e-01
#> [12,] -9.768905e-01
#> [13,] -9.768905e-01
#> [14,] -9.768905e-01
#> [15,] -9.768905e-01
#> [16,] -9.768905e-01
#> [17,] -9.768905e-01
#> [18,] -9.768905e-01
#> [19,] -9.768905e-01
#> [20,] -9.768905e-01
#> [21,] -9.768905e-01
#> [22,] -9.768905e-01
#> [23,] -1.594770e+00
#> [24,] -1.594770e+00
#> [25,] -1.594770e+00
#> [26,] -1.594770e+00
#> [27,] -1.594770e+00
#> [28,] -1.594770e+00
#> [29,] -1.594770e+00
#> [30,] -1.594770e+00
#> [31,] -1.594770e+00
#> [32,] -1.594770e+00
#> [33,] -1.594770e+00
#> [34,] -1.594770e+00
#> [35,] -1.594770e+00
#> [36,] -1.594770e+00
#> [37,] -1.619078e+00
#> [38,] -1.747204e+00
#> [39,] -1.747204e+00
#> [40,] -1.747204e+00
#> [41,] -1.747204e+00
#> [42,] -1.747204e+00
#> [43,] -1.747204e+00
#> [44,] -1.747204e+00
#> [45,] -1.747204e+00
#> [46,] -1.947337e+00
#> [47,] -1.947337e+00
#> [48,] -1.947337e+00
#> [49,] -1.947337e+00
#> [50,] -1.947337e+00
#> [51,] -1.947337e+00
#> [52,] -1.947337e+00
#> [53,] -1.947337e+00
#> [54,] -1.947337e+00
#> [55,] -1.947337e+00
#> [56,] -1.947337e+00
#> [57,] -1.947337e+00
#> [58,] -1.947337e+00
#> [59,] -1.947337e+00
#> [60,] -1.947337e+00
#> [61,] -1.959349e+00
#> [62,] -1.959349e+00
#> [63,] -1.959349e+00
#> [64,] -1.959349e+00
#> [65,] -1.959349e+00
#> [66,] -1.959349e+00
#> [67,] -1.959349e+00
#> [68,] -1.959349e+00
#> [69,] -1.961248e+00
#> [70,] -1.961248e+00
#> [71,] -1.961248e+00
#> [72,] -1.961248e+00
#> [73,] -1.961248e+00
#> [74,] -1.961248e+00
#> [75,] -1.961248e+00
#> [76,] -1.962718e+00
#> [77,] -1.962718e+00
#> [78,] -1.962718e+00
#> [79,] -1.963546e+00
#> [80,] -1.963546e+00
#> [81,] -1.963546e+00
#> [82,] -1.965519e+00
#> [83,] -1.965519e+00
#> [84,] -1.965519e+00
#> [85,] -1.965519e+00
#> [86,] -1.965519e+00
#> [87,] -1.965519e+00
#> 
#> 

#======== examples end =======================