Implementation of the Beetle Swarm Antennae Search (BSAS) algorithm for optimization problems.

You could find more information about BSAS in https://arxiv.org/abs/1807.10470.

BSASoptim(fn, init = NULL, lower = c(-6, 0), upper = c(-1, 2),
  k = 5, d0 = 0.001, d1 = 3, eta_d = 0.95, l0 = 0, l1 = 0,
  eta_l = 0.95, step = 0.8, eta_step = 0.95, n = 200,
  seed = NULL, trace = T, steptol = 0.01, p_min = 0.2,
  p_step = 0.2, n_flag = 2)

Arguments

fn
objective function; function need to be optimized

init
default = NULL, it will generate randomly; Of course, you can specify it.

lower
lower of parameters to be estimated; Default = c(-6,0) because of the test on Michalewicz function of which thelower is c(-6,0); By the way, you should set one of init or lower parameter at least to make the code know the dimensionality of your problem.

upper
upper of parameters; Default = c(-1,2).

k
a positive integer.k is the number of beetles for exploring in every iteration.

d0
a constant to gurantee that sensing length of antennae d doesn't equal to zero. More specifically, $$d^t = \eta_d * d^{t-1} + d_0$$where attenuation coefficient \(\eta_d\) belongs to \([0,1]\)

d1
initial value of antenae length. You can specify it according to your problem scale

eta_d
attenuation coefficient of sensing length of antennae

l0
position jitter factor constant.Default = 0.

l1
initial position jitter factor.Default = 0.$$x = x - step * dir * sign(fn(left) - fn(right)) + l *random(npars)$$

eta_l
attenuation coefficient of jitter factor.$$l^t = \eta_l * l^{t-1} + l_0$$

step
initial step-size of beetle

eta_step
attenuation coefficient of step-size.$$step^t = \eta_step * step^{t-1}$$

n
iterations times

seed
random seed; default = NULL ; The results of BAS depend on random init value and random directions. Therefore, if you set a random seed, for example,seed = 1, the results will remain the same no matter how many times you repeat your experiments.

trace
default = T; trace the process of BAS iteration.

steptol
default = 0.01; Iteration will stop if step-size in current moment is less than steptol.

p_min
a constant belongs to [0,1]. If random generated value is lager than p_min and there are better positions in k beetles than current position, the next position of beetle will be the best position in k beetles. $$x_{best} = argmin(fn(x_i^t))$$ where \(i belongs [1,2,...,k]\) If random generated value is smaller than(<=) p_min and there are better positions in k beetles than current position, the next position of the beetle could be the random position within better positions.

p_step
a constant belongs to [0,1].If no better position in k beetles than current position, there is still a samll probability that step-size doesn't update. If you set a little k, you could set p_step slightly large.

n_flag
an positive integer; default = 2; If step-size doesn't update for successive n_flag times because p_step is larger than random generated value, the step-size will be forced updating. If you set a large p_step, set a small n_flag is suggested.

Value

A list including best beetle position (parameters) and corresponding objective function value.

References

X. Y. Jiang, and S. Li, BAS: beetle antennae search algorithm for optimization problems, arXiv:1710.10724v1.

Examples

#======== examples start =======================
# BSAS application on Michalewicz function
library(rBAS)
mich <- function(x){
   y1 <- -sin(x[1])*(sin((x[1]^2)/pi))^20
   y2 <- -sin(x[2])*(sin((2*x[2]^2)/pi))^20
   return(y1+y2)
}
BSASoptim(fn = mich,
          lower = c(-6,0), upper = c(-1,2),
          seed = 12, n = 100,k=5)

#> ============ Initialization =========== 
#> Iter:  0  xbest:  [ -5.653195 1.63555 ], 
#> fbest=  -0.8377777 
#> ============ Iter  1  start =========== 
#> 5 beetles'objective value : [ -0.03003036 -0.002906397 -0.0301622 -0.0302197 -0.3063353 ] 
#> Position & Objective remain the same -0.8377777 
#> Step-size should reduce
#> Step-size: 0.8 
#> ============ Iter  2  start =========== 
#> 5 beetles'objective value : [ -0.006527109 -1.177867 -0.7724191 -0.03003036 -0.03034871 ] 
#> ***position updates*** 
#> Position/Params updates [ -4.853364 1.619117 ]
#> Corresponding Objective value: -1.177867 
#> ============ Iter  3  start =========== 
#> 5 beetles'objective value : [ -0.3158469 -0.001641267 -0.002149374 -0.01562099 -0.02243852 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.8 
#> ============ Iter  4  start =========== 
#> 5 beetles'objective value : [ -8.546951e-06 -0.0001067939 -8.546951e-06 -8.546951e-06 -2.247376e-07 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.8 
#> ============ Iter  5  start =========== 
#> 5 beetles'objective value : [ -0.01170067 -0.0004296308 -0.0001994871 -8.548614e-06 -0.3690519 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.76 
#> ============ Iter  6  start =========== 
#> 5 beetles'objective value : [ -0.92152 -8.546951e-06 -2.518048e-06 -0.1327225 -1.476056e-05 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.722 
#> ============ Iter  7  start =========== 
#> 5 beetles'objective value : [ -0.000278546 -2.595489e-05 -7.24461e-05 -0.002522549 -8.605598e-06 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.6859 
#> ============ Iter  8  start =========== 
#> 5 beetles'objective value : [ -9.441651e-06 -0.001250998 -1.857947e-05 -2.133368e-05 -5.136194e-05 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.651605 
#> ============ Iter  9  start =========== 
#> 5 beetles'objective value : [ -0.5079644 -8.546951e-06 -0.07081859 -0.0001983879 -0.03069209 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.6190247 
#> ============ Iter  10  start =========== 
#> 5 beetles'objective value : [ -0.0001067129 -0.003560598 -0.9672288 -0.0261916 -0.0004745052 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.5880735 
#> ============ Iter  11  start =========== 
#> 5 beetles'objective value : [ -0.0221035 -0.000585714 -0.7356137 -0.9999931 -0.1119387 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.5586698 
#> ============ Iter  12  start =========== 
#> 5 beetles'objective value : [ -0.963783 -0.9189142 -0.04513856 -0.9253749 -0.001817755 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.5307363 
#> ============ Iter  13  start =========== 
#> 5 beetles'objective value : [ -0.01345954 -0.9999568 -0.0007358565 -0.009769374 -8.546951e-06 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.5041995 
#> ============ Iter  14  start =========== 
#> 5 beetles'objective value : [ -0.9056201 -0.008966723 -0.0003776324 -0.7137564 -0.005153772 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.5041995 
#> ============ Iter  15  start =========== 
#> 5 beetles'objective value : [ -0.00474309 -0.8310443 -0.1121344 -0.00371413 -0.06544858 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.4789896 
#> ============ Iter  16  start =========== 
#> 5 beetles'objective value : [ -0.0006666964 -0.999819 -0.03583166 -0.5107749 -0.0004909553 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.4550401 
#> ============ Iter  17  start =========== 
#> 5 beetles'objective value : [ -0.4121274 -0.0002043546 -0.4324166 -0.008104534 -0.009341122 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.4322881 
#> ============ Iter  18  start =========== 
#> 5 beetles'objective value : [ -0.915839 -0.0533368 -0.998502 -0.9685057 -0.3222134 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.4106737 
#> ============ Iter  19  start =========== 
#> 5 beetles'objective value : [ -0.848857 -0.291086 -0.02135792 -0.3292725 -0.9368057 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.39014 
#> ============ Iter  20  start =========== 
#> 5 beetles'objective value : [ -0.5303926 -0.2164963 -0.001908701 -0.9953929 -0.003366519 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.370633 
#> ============ Iter  21  start =========== 
#> 5 beetles'objective value : [ -0.9932839 -0.187404 -0.9898887 -0.009957931 -0.1966089 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.370633 
#> ============ Iter  22  start =========== 
#> 5 beetles'objective value : [ -0.4732579 -0.6668996 -0.0008740727 -0.9671825 -0.0008225686 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.3521013 
#> ============ Iter  23  start =========== 
#> 5 beetles'objective value : [ -0.85542 -0.2885355 -0.006911496 -0.6808978 -0.004190474 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.3344963 
#> ============ Iter  24  start =========== 
#> 5 beetles'objective value : [ -0.892165 -0.4314826 -0.01081667 -0.09196175 -0.5044806 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.3177715 
#> ============ Iter  25  start =========== 
#> 5 beetles'objective value : [ -0.353504 -0.3932281 -0.9732968 -1.008935 -1.045684 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.3018829 
#> ============ Iter  26  start =========== 
#> 5 beetles'objective value : [ -0.006948546 -0.8658795 -0.7385401 -0.3821867 -0.3486964 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.2867887 
#> ============ Iter  27  start =========== 
#> 5 beetles'objective value : [ -0.7619605 -0.8020957 -0.7926605 -0.01225382 -0.5501577 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.2724493 
#> ============ Iter  28  start =========== 
#> 5 beetles'objective value : [ -1.051591 -0.4703636 -0.5711343 -0.9327761 -0.9088408 ] 
#> Position & Objective remain the same -1.177867 
#> Step-size should reduce
#> Step-size: 0.2588268 
#> ============ Iter  29  start =========== 
#> 5 beetles'objective value : [ -1.207671 -1.250229 -1.030366 -0.3719873 -0.3968342 ] 
#> ***position updates*** 
#> Position/Params updates [ -4.964083 1.385167 ]
#> Corresponding Objective value: -1.250229 
#> ============ Iter  30  start =========== 
#> 5 beetles'objective value : [ -0.5190497 -1.174872 -0.5190604 -1.763478 -0.4966268 ] 
#> ***position updates*** 
#> Position/Params updates [ -5.217324 1.438648 ]
#> Corresponding Objective value: -1.763478 
#> ============ Iter  31  start =========== 
#> 5 beetles'objective value : [ -0.8562412 -1.761961 -0.1092086 -0.9162577 -1.097709 ] 
#> Position & Objective remain the same -1.763478 
#> Step-size should reduce
#> Step-size: 0.2458855 
#> ============ Iter  32  start =========== 
#> 5 beetles'objective value : [ -0.08995107 -0.04164282 -0.9836837 -0.9762342 -0.9978899 ] 
#> Position & Objective remain the same -1.763478 
#> Step-size should reduce
#> Step-size: 0.2335912 
#> ============ Iter  33  start =========== 
#> 5 beetles'objective value : [ -1.579752 -0.2403424 -1.783152 -1.234363 -0.7126726 ] 
#> ***position updates*** 
#> Position/Params updates [ -4.985098 1.46386 ]
#> Corresponding Objective value: -1.783152 
#> ============ Iter  34  start =========== 
#> 5 beetles'objective value : [ -0.7333993 -1.024486 -1.001056 -1.001996 -0.7368599 ] 
#> Position & Objective remain the same -1.783152 
#> Step-size should reduce
#> Step-size: 0.2219117 
#> ============ Iter  35  start =========== 
#> 5 beetles'objective value : [ -0.9175802 -1.047831 -0.8260515 -0.7866107 -1.048697 ] 
#> Position & Objective remain the same -1.783152 
#> Step-size should reduce
#> Step-size: 0.2108161 
#> ============ Iter  36  start =========== 
#> 5 beetles'objective value : [ -1.457275 -1.578228 -1.120181 -0.6594683 -0.5951288 ] 
#> Position & Objective remain the same -1.783152 
#> Step-size should reduce
#> Step-size: 0.2108161 
#> ============ Iter  37  start =========== 
#> 5 beetles'objective value : [ -0.9772363 -0.8132779 -1.561098 -1.591219 -0.9699972 ] 
#> Position & Objective remain the same -1.783152 
#> Step-size should reduce
#> Step-size: 0.2108161 
#> ============ Iter  38  start =========== 
#> 5 beetles'objective value : [ -0.9898467 -0.9650599 -1.11882 -0.9874221 -0.9718968 ] 
#> Position & Objective remain the same -1.783152 
#> Step-size should reduce
#> Step-size: 0.2108161 
#> ============ Iter  39  start =========== 
#> 5 beetles'objective value : [ -0.8230338 -0.9929199 -0.8580985 -0.9233844 -0.8930726 ] 
#> Position & Objective remain the same -1.783152 
#> Step-size should reduce
#> Step-size: 0.2002753 
#> ============ Iter  40  start =========== 
#> 5 beetles'objective value : [ -1.074534 -0.786063 -1.012055 -0.9936264 -1.000344 ] 
#> Position & Objective remain the same -1.783152 
#> Step-size should reduce
#> Step-size: 0.2002753 
#> ============ Iter  41  start =========== 
#> 5 beetles'objective value : [ -1.363023 -0.9907998 -0.9526779 -1.555024 -0.8750123 ] 
#> Position & Objective remain the same -1.783152 
#> Step-size should reduce
#> Step-size: 0.2002753 
#> ============ Iter  42  start =========== 
#> 5 beetles'objective value : [ -1.433687 -0.8819231 -1.086452 -1.145797 -1.17236 ] 
#> Position & Objective remain the same -1.783152 
#> Step-size should reduce
#> Step-size: 0.1902615 
#> ============ Iter  43  start =========== 
#> 5 beetles'objective value : [ -0.7878598 -1.315154 -1.36254 -1.006876 -0.9294633 ] 
#> Position & Objective remain the same -1.783152 
#> Step-size should reduce
#> Step-size: 0.1807484 
#> ============ Iter  44  start =========== 
#> 5 beetles'objective value : [ -1.428016 -1.016422 -1.405206 -1.054458 -1.019376 ] 
#> Position & Objective remain the same -1.783152 
#> Step-size should reduce
#> Step-size: 0.171711 
#> ============ Iter  45  start =========== 
#> 5 beetles'objective value : [ -0.8835519 -1.142698 -1.242856 -1.002362 -1.083458 ] 
#> Position & Objective remain the same -1.783152 
#> Step-size should reduce
#> Step-size: 0.1631255 
#> ============ Iter  46  start =========== 
#> 5 beetles'objective value : [ -1.62254 -1.127685 -1.378717 -0.9740809 -1.188669 ] 
#> Position & Objective remain the same -1.783152 
#> Step-size should reduce
#> Step-size: 0.1549692 
#> ============ Iter  47  start =========== 
#> 5 beetles'objective value : [ -1.349327 -1.361066 -1.13445 -0.8375903 -0.9059747 ] 
#> Position & Objective remain the same -1.783152 
#> Step-size should reduce
#> Step-size: 0.1472207 
#> ============ Iter  48  start =========== 
#> 5 beetles'objective value : [ -1.140708 -1.46122 -1.856251 -1.786802 -1.367827 ] 
#> ***position updates*** 
#> Position/Params updates [ -4.939055 1.603696 ]
#> Corresponding Objective value: -1.856251 
#> ============ Iter  49  start =========== 
#> 5 beetles'objective value : [ -1.048941 -1.202013 -1.371126 -1.258138 -1.118221 ] 
#> Position & Objective remain the same -1.856251 
#> Step-size should reduce
#> Step-size: 0.1398597 
#> ============ Iter  50  start =========== 
#> 5 beetles'objective value : [ -1.184996 -1.429921 -1.049661 -1.053679 -1.231876 ] 
#> Position & Objective remain the same -1.856251 
#> Step-size should reduce
#> Step-size: 0.1328667 
#> ============ Iter  51  start =========== 
#> 5 beetles'objective value : [ -1.322973 -1.533303 -1.221357 -1.101644 -1.544115 ] 
#> Position & Objective remain the same -1.856251 
#> Step-size should reduce
#> Step-size: 0.1262234 
#> ============ Iter  52  start =========== 
#> 5 beetles'objective value : [ -1.634009 -1.653297 -1.701987 -1.701706 -1.05835 ] 
#> Position & Objective remain the same -1.856251 
#> Step-size should reduce
#> Step-size: 0.1262234 
#> ============ Iter  53  start =========== 
#> 5 beetles'objective value : [ -1.678003 -1.299672 -1.092464 -1.20269 -1.299812 ] 
#> Position & Objective remain the same -1.856251 
#> Step-size should reduce
#> Step-size: 0.1199122 
#> ============ Iter  54  start =========== 
#> 5 beetles'objective value : [ -1.336061 -1.465727 -1.733704 -1.609339 -1.323544 ] 
#> Position & Objective remain the same -1.856251 
#> Step-size should reduce
#> Step-size: 0.1139166 
#> ============ Iter  55  start =========== 
#> 5 beetles'objective value : [ -1.716962 -1.755987 -1.36877 -1.425386 -1.114758 ] 
#> Position & Objective remain the same -1.856251 
#> Step-size should reduce
#> Step-size: 0.1082208 
#> ============ Iter  56  start =========== 
#> 5 beetles'objective value : [ -1.205929 -1.176344 -1.782465 -1.790269 -1.142886 ] 
#> Position & Objective remain the same -1.856251 
#> Step-size should reduce
#> Step-size: 0.1028097 
#> ============ Iter  57  start =========== 
#> 5 beetles'objective value : [ -1.801532 -1.706468 -1.15242 -1.729435 -1.327751 ] 
#> Position & Objective remain the same -1.856251 
#> Step-size should reduce
#> Step-size: 0.09766924 
#> ============ Iter  58  start =========== 
#> 5 beetles'objective value : [ -1.77854 -1.341793 -1.774408 -1.414778 -1.630429 ] 
#> Position & Objective remain the same -1.856251 
#> Step-size should reduce
#> Step-size: 0.09278578 
#> ============ Iter  59  start =========== 
#> 5 beetles'objective value : [ -1.810007 -1.704561 -1.770713 -1.500975 -1.840736 ] 
#> Position & Objective remain the same -1.856251 
#> Step-size should reduce
#> Step-size: 0.09278578 
#> ============ Iter  60  start =========== 
#> 5 beetles'objective value : [ -1.362279 -1.492919 -1.373111 -1.449739 -1.83075 ] 
#> Position & Objective remain the same -1.856251 
#> Step-size should reduce
#> Step-size: 0.09278578 
#> ============ Iter  61  start =========== 
#> 5 beetles'objective value : [ -1.702876 -1.491894 -1.31944 -1.853528 -1.852838 ] 
#> Position & Objective remain the same -1.856251 
#> Step-size should reduce
#> Step-size: 0.08814649 
#> ============ Iter  62  start =========== 
#> 5 beetles'objective value : [ -1.661314 -1.861104 -1.637951 -1.312862 -1.864702 ] 
#> ***position updates*** 
#> Position/Params updates [ -4.985146 1.52856 ]
#> Corresponding Objective value: -1.864702 
#> ============ Iter  63  start =========== 
#> 5 beetles'objective value : [ -1.892621 -1.562603 -1.836006 -1.354629 -1.684425 ] 
#> ***position updates*** 
#> Position/Params updates [ -4.959424 1.61287 ]
#> Corresponding Objective value: -1.892621 
#> ============ Iter  64  start =========== 
#> 5 beetles'objective value : [ -1.577913 -1.711298 -1.389379 -1.712214 -1.878743 ] 
#> Position & Objective remain the same -1.892621 
#> Step-size should reduce
#> Step-size: 0.08373916 
#> ============ Iter  65  start =========== 
#> 5 beetles'objective value : [ -1.361471 -1.581051 -1.478069 -1.459619 -1.814304 ] 
#> Position & Objective remain the same -1.892621 
#> Step-size should reduce
#> Step-size: 0.07955221 
#> ============ Iter  66  start =========== 
#> 5 beetles'objective value : [ -1.788375 -1.39565 -1.564254 -1.410108 -1.73147 ] 
#> Position & Objective remain the same -1.892621 
#> Step-size should reduce
#> Step-size: 0.0755746 
#> ============ Iter  67  start =========== 
#> 5 beetles'objective value : [ -1.804923 -1.921501 -1.677714 -1.818033 -1.527951 ] 
#> ***position updates*** 
#> Position/Params updates [ -4.960608 1.537305 ]
#> Corresponding Objective value: -1.921501 
#> ============ Iter  68  start =========== 
#> 5 beetles'objective value : [ -1.744854 -1.89164 -1.743317 -1.717899 -1.735855 ] 
#> Position & Objective remain the same -1.921501 
#> Step-size should reduce
#> Step-size: 0.07179587 
#> ============ Iter  69  start =========== 
#> 5 beetles'objective value : [ -1.531991 -1.604793 -1.632378 -1.842495 -1.909068 ] 
#> Position & Objective remain the same -1.921501 
#> Step-size should reduce
#> Step-size: 0.06820607 
#> ============ Iter  70  start =========== 
#> 5 beetles'objective value : [ -1.554035 -1.599164 -1.884721 -1.560528 -1.682354 ] 
#> Position & Objective remain the same -1.921501 
#> Step-size should reduce
#> Step-size: 0.06820607 
#> ============ Iter  71  start =========== 
#> 5 beetles'objective value : [ -1.891425 -1.812117 -1.583278 -1.801932 -1.586967 ] 
#> Position & Objective remain the same -1.921501 
#> Step-size should reduce
#> Step-size: 0.06479577 
#> ============ Iter  72  start =========== 
#> 5 beetles'objective value : [ -1.724883 -1.925271 -1.752643 -1.565068 -1.695659 ] 
#> ***position updates*** 
#> Position/Params updates [ -4.960541 1.602101 ]
#> Corresponding Objective value: -1.925271 
#> ============ Iter  73  start =========== 
#> 5 beetles'objective value : [ -1.710701 -1.854283 -1.860104 -1.727663 -1.662067 ] 
#> Position & Objective remain the same -1.925271 
#> Step-size should reduce
#> Step-size: 0.06155598 
#> ============ Iter  74  start =========== 
#> 5 beetles'objective value : [ -1.893173 -1.835828 -1.898106 -1.769115 -1.607311 ] 
#> Position & Objective remain the same -1.925271 
#> Step-size should reduce
#> Step-size: 0.05847818 
#> ============ Iter  75  start =========== 
#> 5 beetles'objective value : [ -1.678783 -1.849819 -1.616488 -1.757929 -1.773269 ] 
#> Position & Objective remain the same -1.925271 
#> Step-size should reduce
#> Step-size: 0.05555427 
#> ============ Iter  76  start =========== 
#> 5 beetles'objective value : [ -1.916067 -1.942968 -1.89417 -1.90009 -1.647402 ] 
#> ***position updates*** 
#> Position/Params updates [ -4.961789 1.54656 ]
#> Corresponding Objective value: -1.942968 
#> ============ Iter  77  start =========== 
#> 5 beetles'objective value : [ -1.910065 -1.908232 -1.887382 -1.895629 -1.7799 ] 
#> Position & Objective remain the same -1.942968 
#> Step-size should reduce
#> Step-size: 0.05277656 
#> ============ Iter  78  start =========== 
#> 5 beetles'objective value : [ -1.839942 -1.809385 -1.803773 -1.801086 -1.813192 ] 
#> Position & Objective remain the same -1.942968 
#> Step-size should reduce
#> Step-size: 0.05277656 
#> ============ Iter  79  start =========== 
#> 5 beetles'objective value : [ -1.912614 -1.918195 -1.897647 -1.909026 -1.750696 ] 
#> Position & Objective remain the same -1.942968 
#> Step-size should reduce
#> Step-size: 0.05013773 
#> ============ Iter  80  start =========== 
#> 5 beetles'objective value : [ -1.840261 -1.842128 -1.92174 -1.891417 -1.774773 ] 
#> Position & Objective remain the same -1.942968 
#> Step-size should reduce
#> Step-size: 0.04763084 
#> ============ Iter  81  start =========== 
#> 5 beetles'objective value : [ -1.895623 -1.884725 -1.761456 -1.754811 -1.797292 ] 
#> Position & Objective remain the same -1.942968 
#> Step-size should reduce
#> Step-size: 0.04763084 
#> ============ Iter  82  start =========== 
#> 5 beetles'objective value : [ -1.85057 -1.946033 -1.76825 -1.94308 -1.902533 ] 
#> ***position updates*** 
#> Position/Params updates [ -4.960885 1.594183 ]
#> Corresponding Objective value: -1.946033 
#> ============ Iter  83  start =========== 
#> 5 beetles'objective value : [ -1.936055 -1.788361 -1.939013 -1.820656 -1.938257 ] 
#> Position & Objective remain the same -1.946033 
#> Step-size should reduce
#> Step-size: 0.04763084 
#> ============ Iter  84  start =========== 
#> 5 beetles'objective value : [ -1.832534 -1.863397 -1.941322 -1.900547 -1.779116 ] 
#> Position & Objective remain the same -1.946033 
#> Step-size should reduce
#> Step-size: 0.0452493 
#> ============ Iter  85  start =========== 
#> 5 beetles'objective value : [ -1.906501 -1.830391 -1.86918 -1.933526 -1.837117 ] 
#> Position & Objective remain the same -1.946033 
#> Step-size should reduce
#> Step-size: 0.0452493 
#> ============ Iter  86  start =========== 
#> 5 beetles'objective value : [ -1.793507 -1.949056 -1.910671 -1.8596 -1.858055 ] 
#> ***position updates*** 
#> Position/Params updates [ -4.969681 1.549797 ]
#> Corresponding Objective value: -1.949056 
#> ============ Iter  87  start =========== 
#> 5 beetles'objective value : [ -1.908504 -1.917964 -1.901125 -1.799799 -1.913223 ] 
#> Position & Objective remain the same -1.949056 
#> Step-size should reduce
#> Step-size: 0.04298684 
#> ============ Iter  88  start =========== 
#> 5 beetles'objective value : [ -1.919931 -1.875032 -1.832292 -1.940719 -1.918314 ] 
#> Position & Objective remain the same -1.949056 
#> Step-size should reduce
#> Step-size: 0.04083749 
#> ============ Iter  89  start =========== 
#> 5 beetles'objective value : [ -1.889609 -1.850674 -1.904524 -1.921623 -1.910445 ] 
#> Position & Objective remain the same -1.949056 
#> Step-size should reduce
#> Step-size: 0.03879562 
#> ============ Iter  90  start =========== 
#> 5 beetles'objective value : [ -1.895597 -1.950848 -1.852005 -1.843603 -1.854016 ] 
#> ***position updates*** 
#> Position/Params updates [ -4.957656 1.586682 ]
#> Corresponding Objective value: -1.950848 
#> ============ Iter  91  start =========== 
#> 5 beetles'objective value : [ -1.874846 -1.868995 -1.927775 -1.857075 -1.925979 ] 
#> Position & Objective remain the same -1.950848 
#> Step-size should reduce
#> Step-size: 0.03685584 
#> ============ Iter  92  start =========== 
#> 5 beetles'objective value : [ -1.909348 -1.94249 -1.912683 -1.949768 -1.901144 ] 
#> Position & Objective remain the same -1.950848 
#> Step-size should reduce
#> Step-size: 0.03501305 
#> ============ Iter  93  start =========== 
#> 5 beetles'objective value : [ -1.943475 -1.877855 -1.881201 -1.907147 -1.906859 ] 
#> Position & Objective remain the same -1.950848 
#> Step-size should reduce
#> Step-size: 0.03326239 
#> ============ Iter  94  start =========== 
#> 5 beetles'objective value : [ -1.957294 -1.914998 -1.883728 -1.915559 -1.885753 ] 
#> ***position updates*** 
#> Position/Params updates [ -4.97003 1.555807 ]
#> Corresponding Objective value: -1.957294 
#> ============ Iter  95  start =========== 
#> 5 beetles'objective value : [ -1.894761 -1.917257 -1.864796 -1.931012 -1.948813 ] 
#> Position & Objective remain the same -1.957294 
#> Step-size should reduce
#> Step-size: 0.03159928 
#> ============ Iter  96  start =========== 
#> 5 beetles'objective value : [ -1.912611 -1.899107 -1.952543 -1.88334 -1.911616 ] 
#> Position & Objective remain the same -1.957294 
#> Step-size should reduce
#> Step-size: 0.03001931 
#> ============ Iter  97  start =========== 
#> 5 beetles'objective value : [ -1.959051 -1.954945 -1.942496 -1.921735 -1.912499 ] 
#> ***position updates*** 
#> Position/Params updates [ -4.964421 1.585297 ]
#> Corresponding Objective value: -1.959051 
#> ============ Iter  98  start =========== 
#> 5 beetles'objective value : [ -1.932917 -1.957433 -1.897393 -1.9124 -1.885254 ] 
#> Position & Objective remain the same -1.959051 
#> Step-size should reduce
#> Step-size: 0.02851835 
#> ============ Iter  99  start =========== 
#> 5 beetles'objective value : [ -1.95599 -1.915233 -1.913963 -1.928009 -1.893062 ] 
#> Position & Objective remain the same -1.959051 
#> Step-size should reduce
#> Step-size: 0.02709243 
#> ============ Iter  100  start =========== 
#> 5 beetles'objective value : [ -1.94024 -1.89545 -1.901827 -1.961165 -1.94011 ] 
#> ***position updates*** 
#> Position/Params updates [ -4.964138 1.558206 ]
#> Corresponding Objective value: -1.961165 

#> $par
#> [1] -4.964138  1.558206
#> 
#> $value
#> [1] -1.961165
#> 
#> $df
#> $df$x
#>             [,1]     [,2]
#>   [1,] -5.653195 1.635550
#>   [2,] -5.653195 1.635550
#>   [3,] -4.853364 1.619117
#>   [4,] -4.853364 1.619117
#>   [5,] -4.853364 1.619117
#>   [6,] -4.853364 1.619117
#>   [7,] -4.853364 1.619117
#>   [8,] -4.853364 1.619117
#>   [9,] -4.853364 1.619117
#>  [10,] -4.853364 1.619117
#>  [11,] -4.853364 1.619117
#>  [12,] -4.853364 1.619117
#>  [13,] -4.853364 1.619117
#>  [14,] -4.853364 1.619117
#>  [15,] -4.853364 1.619117
#>  [16,] -4.853364 1.619117
#>  [17,] -4.853364 1.619117
#>  [18,] -4.853364 1.619117
#>  [19,] -4.853364 1.619117
#>  [20,] -4.853364 1.619117
#>  [21,] -4.853364 1.619117
#>  [22,] -4.853364 1.619117
#>  [23,] -4.853364 1.619117
#>  [24,] -4.853364 1.619117
#>  [25,] -4.853364 1.619117
#>  [26,] -4.853364 1.619117
#>  [27,] -4.853364 1.619117
#>  [28,] -4.853364 1.619117
#>  [29,] -4.853364 1.619117
#>  [30,] -4.964083 1.385167
#>  [31,] -5.217324 1.438648
#>  [32,] -5.217324 1.438648
#>  [33,] -5.217324 1.438648
#>  [34,] -4.985098 1.463860
#>  [35,] -4.985098 1.463860
#>  [36,] -4.985098 1.463860
#>  [37,] -4.985098 1.463860
#>  [38,] -4.985098 1.463860
#>  [39,] -4.985098 1.463860
#>  [40,] -4.985098 1.463860
#>  [41,] -4.985098 1.463860
#>  [42,] -4.985098 1.463860
#>  [43,] -4.985098 1.463860
#>  [44,] -4.985098 1.463860
#>  [45,] -4.985098 1.463860
#>  [46,] -4.985098 1.463860
#>  [47,] -4.985098 1.463860
#>  [48,] -4.985098 1.463860
#>  [49,] -4.939055 1.603696
#>  [50,] -4.939055 1.603696
#>  [51,] -4.939055 1.603696
#>  [52,] -4.939055 1.603696
#>  [53,] -4.939055 1.603696
#>  [54,] -4.939055 1.603696
#>  [55,] -4.939055 1.603696
#>  [56,] -4.939055 1.603696
#>  [57,] -4.939055 1.603696
#>  [58,] -4.939055 1.603696
#>  [59,] -4.939055 1.603696
#>  [60,] -4.939055 1.603696
#>  [61,] -4.939055 1.603696
#>  [62,] -4.939055 1.603696
#>  [63,] -4.985146 1.528560
#>  [64,] -4.959424 1.612870
#>  [65,] -4.959424 1.612870
#>  [66,] -4.959424 1.612870
#>  [67,] -4.959424 1.612870
#>  [68,] -4.960608 1.537305
#>  [69,] -4.960608 1.537305
#>  [70,] -4.960608 1.537305
#>  [71,] -4.960608 1.537305
#>  [72,] -4.960608 1.537305
#>  [73,] -4.960541 1.602101
#>  [74,] -4.960541 1.602101
#>  [75,] -4.960541 1.602101
#>  [76,] -4.960541 1.602101
#>  [77,] -4.961789 1.546560
#>  [78,] -4.961789 1.546560
#>  [79,] -4.961789 1.546560
#>  [80,] -4.961789 1.546560
#>  [81,] -4.961789 1.546560
#>  [82,] -4.961789 1.546560
#>  [83,] -4.960885 1.594183
#>  [84,] -4.960885 1.594183
#>  [85,] -4.960885 1.594183
#>  [86,] -4.960885 1.594183
#>  [87,] -4.969681 1.549797
#>  [88,] -4.969681 1.549797
#>  [89,] -4.969681 1.549797
#>  [90,] -4.969681 1.549797
#>  [91,] -4.957656 1.586682
#>  [92,] -4.957656 1.586682
#>  [93,] -4.957656 1.586682
#>  [94,] -4.957656 1.586682
#>  [95,] -4.970030 1.555807
#>  [96,] -4.970030 1.555807
#>  [97,] -4.970030 1.555807
#>  [98,] -4.964421 1.585297
#>  [99,] -4.964421 1.585297
#> [100,] -4.964421 1.585297
#> [101,] -4.964138 1.558206
#> 
#> $df$f
#>              [,1]
#>   [1,] -0.8377777
#>   [2,] -0.8377777
#>   [3,] -1.1778672
#>   [4,] -1.1778672
#>   [5,] -1.1778672
#>   [6,] -1.1778672
#>   [7,] -1.1778672
#>   [8,] -1.1778672
#>   [9,] -1.1778672
#>  [10,] -1.1778672
#>  [11,] -1.1778672
#>  [12,] -1.1778672
#>  [13,] -1.1778672
#>  [14,] -1.1778672
#>  [15,] -1.1778672
#>  [16,] -1.1778672
#>  [17,] -1.1778672
#>  [18,] -1.1778672
#>  [19,] -1.1778672
#>  [20,] -1.1778672
#>  [21,] -1.1778672
#>  [22,] -1.1778672
#>  [23,] -1.1778672
#>  [24,] -1.1778672
#>  [25,] -1.1778672
#>  [26,] -1.1778672
#>  [27,] -1.1778672
#>  [28,] -1.1778672
#>  [29,] -1.1778672
#>  [30,] -1.2502286
#>  [31,] -1.7634783
#>  [32,] -1.7634783
#>  [33,] -1.7634783
#>  [34,] -1.7831519
#>  [35,] -1.7831519
#>  [36,] -1.7831519
#>  [37,] -1.7831519
#>  [38,] -1.7831519
#>  [39,] -1.7831519
#>  [40,] -1.7831519
#>  [41,] -1.7831519
#>  [42,] -1.7831519
#>  [43,] -1.7831519
#>  [44,] -1.7831519
#>  [45,] -1.7831519
#>  [46,] -1.7831519
#>  [47,] -1.7831519
#>  [48,] -1.7831519
#>  [49,] -1.8562510
#>  [50,] -1.8562510
#>  [51,] -1.8562510
#>  [52,] -1.8562510
#>  [53,] -1.8562510
#>  [54,] -1.8562510
#>  [55,] -1.8562510
#>  [56,] -1.8562510
#>  [57,] -1.8562510
#>  [58,] -1.8562510
#>  [59,] -1.8562510
#>  [60,] -1.8562510
#>  [61,] -1.8562510
#>  [62,] -1.8562510
#>  [63,] -1.8647022
#>  [64,] -1.8926215
#>  [65,] -1.8926215
#>  [66,] -1.8926215
#>  [67,] -1.8926215
#>  [68,] -1.9215009
#>  [69,] -1.9215009
#>  [70,] -1.9215009
#>  [71,] -1.9215009
#>  [72,] -1.9215009
#>  [73,] -1.9252713
#>  [74,] -1.9252713
#>  [75,] -1.9252713
#>  [76,] -1.9252713
#>  [77,] -1.9429676
#>  [78,] -1.9429676
#>  [79,] -1.9429676
#>  [80,] -1.9429676
#>  [81,] -1.9429676
#>  [82,] -1.9429676
#>  [83,] -1.9460331
#>  [84,] -1.9460331
#>  [85,] -1.9460331
#>  [86,] -1.9460331
#>  [87,] -1.9490559
#>  [88,] -1.9490559
#>  [89,] -1.9490559
#>  [90,] -1.9490559
#>  [91,] -1.9508484
#>  [92,] -1.9508484
#>  [93,] -1.9508484
#>  [94,] -1.9508484
#>  [95,] -1.9572943
#>  [96,] -1.9572943
#>  [97,] -1.9572943
#>  [98,] -1.9590508
#>  [99,] -1.9590508
#> [100,] -1.9590508
#> [101,] -1.9611652
#> 
#> 

#======== examples end =======================